Move inMemorySender to tests
